Fixed my idiotic mistake
Moved plugin.yml into /src/
Replaced deprecation
Fixed inventoryview problems
Removed unused variables
Made HolographicDisplays a dependency (Could be a mistake on my part)
