+++ .classpath
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/Spigot"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/API"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
