--- Keycrates/bin/config.yml
@@ -1,3 +1,0 @@
-#    --------------
-#   |  KeyCrates  |               
-#    --------------
